+++ CHANGES.rst
@@ -2,6 +2,7 @@
 {next}
 ------
 
+* Changed: Support docutils math directive instead of rst2pdf's (Issue #722)
 * Fixed: HTTPS images using image directive now works with Python 3 (Issue #780)
 * Changed: Update tested dependencies to include ReportLab 3.5.23 (Issue #779)
 * Fixed: Internal links work when using ReportLab 3.5.19+ (Issue #772)